Broke up schemas into independent models
